producer/requirements.txt
finnhub-python==2.4.26
websocket-client==1.9.0
python-dotenv==1.2.1
kafka-python==2.3.0
python-snappy==0.7.3